+++ packages/client/src/client.ts
@@ -26,6 +26,14 @@ export type UrlMatcher =
   | RegExp
   | ((url: URL) => boolean | Promise<boolean>);
 
+/** Possible values for route type */
+export type RouteType = "server-only" | "client-only" | "both";
+export const RouteType = {
+  ServerOnly: "server-only",
+  ClientOnly: "client-only",
+  Both: "both",
+} as const;
+
 /** Options when configuring a route */
 export interface RouteOptions {
   /**
@@ -35,6 +43,16 @@ export interface RouteOptions {
    * When `undefined`, the route handler will be run indefinitely.
    */
   times?: number;
+  /**
+   * Defines the behaviour of the route handler.
+   *
+   * - `server-only` - The route handler will only be called if the request is executed on the server.
+   * - `client-only` - The route handler will only be called if the request is executed on the client.
+   * - `both` - The route handler will be called regardless of whether the request is executed on the server or client.
+   *
+   * @default "both"
+   */
+  type?: RouteType;
 }
 
 type RouteMeta = RouteOptions & {
@@ -48,6 +66,16 @@ export interface WaitForRequestOptions {
    * @default {@link DefaultWaitForRequestTimeout}
    */
   timeout?: number;
+  /**
+   * Determines where to expect the request to be executed from
+   *
+   * - `server-only` - The request will only be received if it originated from the server.
+   * - `client-only` - The request will only be received if it originated from the client.
+   * - `both` - The request will be received regardless of whether it originated from the server or client. The request returned is the first request received.
+   *
+   * @default "both"
+   */
+  type?: RouteType;
 }
 
 /**
@@ -82,7 +110,17 @@ export interface ResponseData {
 }
 
 /**
- * Client used to interface with the WebSocket server for mocking server-side network requests.
+ * External handlers should not handle explicit fallbacks as this is used
+ * internally to continue to the next handler
+ */
+export type ExternalRouteHandlerRouteResponse = Exclude<
+  RouteResponse,
+  { type: "fallback" }
+>;
+
+/**
+ * Client used to interface with the WebSocket server for mocking server-side network requests. And
+ * optionally integrating for mocking client-side network requests.
  *
  * @hideconstructor
  */
@@ -96,6 +134,15 @@ export class Client {
     MessageType,
     Set<(message: ParsedMessage) => void | Promise<void>>
   > = new Map();
+  /** Tracks whether an external client side route handler is attached */
+  private externalClientSideRouteHandlerAttached = false;
+  /** Callback handlers for waiting on client-side network requests */
+  private clientWaitForRequestHandlers: Set<
+    (request: Request) => Promise<void>
+  > = new Set();
+
+  /** Convenient access to route types */
+  public readonly RouteType = RouteType;
 
   /** Registered route handlers to handle outgoing network requests on the server */
   private routeHandlers: Map<
@@ -226,31 +273,93 @@ export class Client {
    */
   async waitForRequest(
     urlMatcher: UrlMatcher,
-    {
-      timeout: timeoutDuration = DefaultWaitForRequestTimeout,
-    }: WaitForRequestOptions = {},
+    options: WaitForRequestOptions = {},
   ): Promise<Request> {
+    const {
+      timeout: timeoutDuration = DefaultWaitForRequestTimeout,
+      type = RouteType.Both,
+    } = options;
     return new Promise<Request>((resolve, reject) => {
       const timeout = setTimeout(() => {
         reject(new Error("Timed out waiting for request"));
       }, timeoutDuration);
 
-      const onRequest = async (
+      const client = this;
+      async function onServerRequest(
         message: ParsedMessageType<MessageTypes["REQUEST"]>,
-      ) => {
-        const urlMatch = await this.doesUrlMatch(
+      ) {
+        const urlMatch = await client.doesUrlMatch(
           urlMatcher,
           new URL(message.payload.request.url),
         );
         if (urlMatch) {
-          this.off(MessageType.REQUEST, onRequest);
+          unregisterHandlers();
           clearTimeout(timeout);
-          resolve(this.getRequestObjectFromRequestMessage(message));
+          resolve(client.getRequestObjectFromRequestMessage(message));
         }
-      };
+      }
 
-      this.on(MessageType.REQUEST, onRequest);
+      async function onClientRequest(request: Request) {
+        const urlMatch = await client.doesUrlMatch(
+          urlMatcher,
+          new URL(request.url),
+        );
+        if (urlMatch) {
+          unregisterHandlers();
+          clearTimeout(timeout);
+          resolve(request);
+        }
+      }
+
+      function unregisterHandlers() {
+        client.off(MessageType.REQUEST, onServerRequest);
+        client.clientWaitForRequestHandlers.delete(onClientRequest);
+      }
+
+      if (this.shouldHandleRouteType(type, "server")) {
+        this.on(MessageType.REQUEST, onServerRequest);
+      }
+
+      if (this.shouldHandleRouteType(type, "client")) {
+        this.clientWaitForRequestHandlers.add(onClientRequest);
+      }
     });
+  }
+
+  /**
+   * Increments the call count for a route handler. Function to be called
+   * whenever a route handler is executed, irregardless of the result.
+   */
+  private incrementRouteHandlerCallCount = (
+    routeHandlerId: string,
+    routeMeta: RouteMeta,
+  ) => {
+    routeMeta.calls++;
+    if (routeMeta.times === routeMeta.calls) {
+      this.unroute(routeHandlerId);
+    }
+  };
+
+  /**
+   * Deduce whether or not a handler should be executed based on the route metadata and handler type.
+   *
+   * @param routeMeta - The metadata & options of the route.
+   * @param handlerType - The type of the handler wanting to execute the route handler.
+   * @returns Whether or not the handler should be executed.
+   */
+  private shouldHandleRouteType(
+    routeType: RouteType | undefined,
+    handlerType: "server" | "client",
+  ) {
+    const type = routeType ?? RouteType.Both;
+    switch (handlerType) {
+      case "server":
+        return type === "server-only" || type === "both";
+      case "client":
+        return type === "client-only" || type === "both";
+      default:
+        throw new Error(`Invalid handler type: ${handlerType}`);
+    }
   }
 
   /**
@@ -275,6 +384,8 @@ export class Client {
       .routeHandlers) {
       // Skip the handler if the URL does not match
       if (!(await this.doesUrlMatch(urlMatcher, url))) continue;
+      // Skip the handler if it should not be handled by the server
+      if (!this.shouldHandleRouteType(routeMeta.type, "server")) continue;
 
       // Execute the handler
       const routeResponse = await handler(route);
@@ -302,13 +413,10 @@ export class Client {
           break;
       }
 
-      // If we have responded check if the route needs to be unregistered then exit
-      if (responded) {
-        routeMeta.calls++;
-        if (routeMeta.times === routeMeta.calls) {
-          this.unroute(routeHandlerId);
-        }
+      this.incrementRouteHandlerCallCount(routeHandlerId, routeMeta);
 
+      // If we have responded then exit
+      if (responded) {
         return;
       }
     }
@@ -467,6 +575,70 @@ export class Client {
    */
   unrouteAll() {
     this.routeHandlers.clear();
+  }
+
+  /**
+   * Allows an external handler to be injected into the route handling process for dealing with
+   * client-side request interception.
+   *
+   * @internal
+   *
+   * @param options - options for the external route handler
+   * @param options.extractRequest - a function that extracts a request from the arguments passed to the external handler
+   * @param options.handleResult - a function that transforms the internal result to the external handlers expected result
+   * @returns a function that acts as a handler for the external route handler
+   */
+  attachExternalClientSideRouteHandler<
+    TCallbackArgs extends unknown[],
+    TResult,
+  >({
+    extractRequest,
+    handleResult,
+  }: {
+    extractRequest: (...args: TCallbackArgs) => Request | Promise<Request>;
+    handleResult: (
+      response: ExternalRouteHandlerRouteResponse | undefined,
+      ...args: TCallbackArgs
+    ) => TResult | undefined | Promise<TResult | undefined>;
+  }): (...args: TCallbackArgs) => Promise<TResult | undefined> {
+    if (this.externalClientSideRouteHandlerAttached) {
+      logger.warn(
+        "External client side route handler already attached. Are you sure you want to call attachExternalClientSideRouteHandler? attachExternalClientSideRouteHandler was only intended to be called once per client instance.",
+      );
+    }
+
+    this.externalClientSideRouteHandlerAttached = true;
+    return async (...args: TCallbackArgs) => {
+      const request = await extractRequest(...args);
+      const url = new URL(request.url);
+      const route = new Route(request);
+
+      for (const awaitingRequestHandler of this.clientWaitForRequestHandlers) {
+        await awaitingRequestHandler(request);
+      }
+
+      for (const [routeHandlerId, [urlMatcher, handler, routeMeta]] of this
+        .routeHandlers) {
+        // Skip the handler if the URL does not match
+        if (!(await this.doesUrlMatch(urlMatcher, url))) continue;
+        // Skip the handler if it should not be handled by the client
+        if (!this.shouldHandleRouteType(routeMeta.type, "client")) continue;
+
+        // Execute the handler
+        const routeResponse = await handler(route);
+        let finalResponse: TResult | undefined;
+        let handledRoute = false;
+        if (routeResponse.type !== "fallback") {
+          handledRoute = true;
+          finalResponse = await handleResult(routeResponse, ...args);
+        }
+
+        this.incrementRouteHandlerCallCount(routeHandlerId, routeMeta);
+        if (handledRoute) return finalResponse;
+      }
+
+      await handleResult(undefined, ...args);
+    };
   }
 
   /**