+++ packages/react-router/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mocky-balboa/react-router
 
+## 1.1.2
+
+### Patch Changes
+
+- [d326511](https://github.com/mocky-balboa/mocky-balboa/commit/d3265110ad1c72af09ef2f85cf543df2d5a5bad2): Documentation updates
+- Updated dependencies [[d326511](https://github.com/mocky-balboa/mocky-balboa/commit/d3265110ad1c72af09ef2f85cf543df2d5a5bad2)]
+  - [@mocky-balboa/cli-utils@1.0.3](https://github.com/mocky-balboa/mocky-balboa/releases/tag/%40mocky-balboa%2Fcli-utils%401.0.3)
+  - [@mocky-balboa/logger@1.0.7](https://github.com/mocky-balboa/mocky-balboa/releases/tag/%40mocky-balboa%2Flogger%401.0.7)
+  - [@mocky-balboa/vite@1.0.3](https://github.com/mocky-balboa/mocky-balboa/releases/tag/%40mocky-balboa%2Fvite%401.0.3)
+
 ## 1.1.1
 
 ### Patch Changes